requirements.txt
streamlit
pandas
numpy
faiss-cpu
duckdb
sentence-transformers
openai
scikit-learn
python-dotenv